agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Serial-to-MQTT bridge for Whacamole game"
readme = "README.md"
license-file = "../LICENSE"
authors = [{ name = "darragh0" }]
requires-python = ">=3.13"
dependencies = ["pyserial>=3.5", "rich-argparse>=1.7.2"]

[project.scripts]
agent = "agent.__main__:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
